style: 코드 리뷰 및 디자인 리뷰 순서 변경
